refactoring button Logout
